🔨 Switch Profile/Navigation to use useOvermind
