+++ packages/app/src/app/pages/Profile/index.tsx
@@ -12,7 +12,7 @@ import { NotFound } from 'app/pages/common/NotFound';
 
 import { Container, Content, Margin } from './elements';
 import Header from './Header';
-import Navigation from './Navigation';
+import { Navigation } from './Navigation';
 import { Sandboxes } from './Sandboxes';
 import { Showcase } from './Showcase';
 
@@ -54,11 +54,7 @@ export const Profile: FunctionComponent<Props> = ({
 
       <Content>
         <MaxWidth>
-          <Navigation
-            likeCount={user.givenLikeCount}
-            sandboxCount={user.sandboxCount}
-            username={user.username}
-          />
+          <Navigation />
         </MaxWidth>
       </Content>
